+++ src/from-promise.ts
@@ -17,7 +17,7 @@ export type IPendingPromise = {
     readonly reason: any;
 }
 
-export type IResolvedPromise<T> = {
+export type IFulfilledPromise<T> = {
     readonly state: "fulfilled";
     readonly value: T;
 }
@@ -27,7 +27,7 @@ export type IRejectedPromise = {
     readonly value: any;
 }
 
-export type IPromiseBasedObservable<T> = IBasePromiseBasedObservable<T> & (IPendingPromise | IResolvedPromise<T> | IRejectedPromise)
+export type IPromiseBasedObservable<T> = IBasePromiseBasedObservable<T> & (IPendingPromise | IFulfilledPromise<T> | IRejectedPromise)
 
 class PromiseBasedObservable<T> {
     @observable.ref value: T = undefined;
@@ -96,15 +96,36 @@ class PromiseBasedObservable<T> {
  * Note that the status strings are available as constants:
  * `mobxUtils.PENDING`, `mobxUtils.REJECTED`, `mobxUtil.FULFILLED`
  *
+ * For testing, promises can be created immediatly in a certain state using
+ * `fromPromise.reject(reason)` or `fromPromise.resolve(reason)`
+ *
  * @param {IThenable<T>} promise The promise which will be observed
  * @param {T} [initialValue=undefined] Optional predefined initial value
  * @returns {IPromiseBasedObservable<T>}
  */
-export function fromPromise<T>(promise: PromiseLike<T>): IPromiseBasedObservable<T> {
+export const fromPromise: {
+    <T>(promise: PromiseLike<T>): IPromiseBasedObservable<T>;
+    reject<T>(reason: any): IRejectedPromise & IBasePromiseBasedObservable<T>;
+    resolve<T>(value?: T): IFulfilledPromise<T> & IBasePromiseBasedObservable<T>;
+} = function(promise: any) {
     invariant(arguments.length === 1, "fromPromise expects exactly one argument");
     invariant(typeof promise === "object" && promise && typeof promise.then === "function", "Please pass a promise to fromPromise");
     return new PromiseBasedObservable(promise) as any;
-}
+} as any;
+
+fromPromise.reject = action("fromPromise.reject", function(reason: any) {
+    const p: any = fromPromise(Promise.reject(reason));
+    p.state = REJECTED;
+    p.value = reason;
+    return p;
+}) as any;
+
+fromPromise.resolve = action("fromPromise.resolve", function(value: any = undefined) {
+    const p: any = fromPromise(Promise.resolve(value));
+    p.state = FULFILLED;
+    p.value = value;
+    return p;
+}) as any;
 
 /**
   * Returns true if the provided value is a promise-based observable.